FREESCAPE: added more sounds for dark and eclipse in zx

+++ engines/freescape/sound.h
@@ -36,7 +36,7 @@ struct soundFx {
 struct soundUnitZX {
 	uint16 freqTimesSeconds;
 	uint16 tStates;
-	uint32 multiplier;
+	float multiplier;
 };
 
 struct soundSpeakerFx {
